feat: Add ExampleController and TestOptions
Add a new file `ExampleController.ts` in the `controllers` folder. This file contains the implementation of the `ExampleController` class, which extends the `Controller` class from the `lakutata/com/entrypoint` module. The `ExampleController` class has a single method `test` that takes an input of type `ActionPattern<TestOptions>` and returns a `Promise<number>`. This method is decorated with `CLIAction`, `HTTPAction`, and `ServiceAction` decorators.

Add a new file `TestOptions.ts` in the `options` folder. This file contains the implementation of the `TestOptions` class, which extends the `DTO` class from the `lakutata` module. The `TestOptions` class has a single property `timestamp` of type `number`, which is decorated with `Expect` decorator.

Update the `Config.ts` file to import and include the `ExampleController` class in the `controllers` array of the `BuildEntrypoints` function.

These changes add the `ExampleController` and `TestOptions` classes to the project and configure the `Config` function to include the `ExampleController` in the application options.

Note: The import statement for the `ExampleController` class in the `Config.ts` file was commented out before this change.